Factored out delineation between floating-point and integer instructions inside instruction set
